proof of concept world to screen working

+++ AC_Cheat/src/entity.h
@@ -19,8 +19,8 @@ public:
 	int entHealth; //0x00EC
 	char pad_00F0[273]; //0x00F0
 	char entName[16] = ""; //0x0204
-	int screenX;
-	int screenY;
+	float screenX;
+	float screenY;
 	void print(entity ent) const {
 		std::cout << "Name: " << ent.entName << std::endl;
 		std::cout << std::dec<<   "Health: " << ent.entHealth << std::endl;
